refactor: remove debugging console calls
`console.log` and similar shouldn't be included within an agnostic library, as they are for debugging purposes. This is especially unnecessary for errors since they can be caught by the client of this library. If such logs are desired, check out a module like [debug](https://github.com/visionmedia/debug).