Remove unused parameters in typescript test file
Unfortunately, you can't ignore this compiler option like a lint option. See [here](https://github.com/Microsoft/TypeScript/issues/11051).
While it would be nice to prove out that the access option defintion is working, there's a lot of things
that aren't included in this test file anyways so it needs to be revisited.

This also switches the hapi key to something that actually works.

This PR is the completion of: https://github.com/brainflake/node-hubspot/pull/97