Correct update customer test
Why:

It looks like the before call is happening in the incorrect place. This
means that the test never gets a valid customer to try and update.

This PR:

Moves the `before` call outside of `it`. The test then began to fail due
to the email not being unique so this change also adds the current
timestamp to the updated email.

This change also removes `console.log` calls.