Add clification for usuage of timelines
Why:

It's not immediately clear that there is one function we expect most
users will be interested in for the timelines api.

This PR:

Adds comments to the usage in the README to clarify that three of the
timelines functions are for setup and one of them is the main usage.