Add test coverage for oauth
Why:

We'd like to have test coverage to sanity check any changes that we
make to the library.

This change addresses the need by:

Adding test coverage for getAccessToken and refreshAccessToken.

NOTE: Since oauth requires a server to do the handshake dance with
HubSpot these tests will not pass with NOCK_OFF=true. We therefore mark
the tests as pending when NOCK_OFF=true.